tweet-comments-service/requirements.txt
psycopg2==2.9.1
grpcio==1.41.0
grpcio-tools==1.41.0
sqlalchemy==1.4.25